requirements.txt
annotated-types==0.7.0
anyio==4.8.0
black==25.1.0
certifi==2025.1.31
click==8.1.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
ollama==0.4.7
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
rich==13.9.4
sniffio==1.3.1
typing_extensions==4.12.2
wcwidth==0.2.13
PyYAML==6.0.1
